async queue is now bigger, replaced offset structure with pointer maths (fixing different struct sizes/offsets in debug builds)

+++ src/cpp/mod/NetworkStuff.cpp
@@ -69,8 +69,7 @@ namespace Network{
 
 			HINTERNET hRequest = HttpOpenRequest(hConnect, method == INetPost ? "POST" : "GET",
 				(script+(method==INetGet?std::string("?"+params):std::string(""))).c_str(), NULL, NULL, accept, 0, 1);
-			BOOL res = HttpSendRequest(hRequest, headers, strlen(headers), (void*)form, strlen(form));
-
+			BOOL res = HttpSendRequest(hRequest, headers, (DWORD)strlen(headers), (void*)form, (DWORD)strlen(form));
 			std::stringstream ss;
 			if (res) {
